% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt.R
\name{adapt}
\alias{adapt}
\title{Adapt Spanish (and Catalan) encoding when using UTF8.}
\usage{
adapt(data, tolower = FALSE)
}
\arguments{
\item{data}{Character vector or data frame. If data frame, converts
all columns of class character.}

\item{tolower}{If TRUE, returns lower-case. Logical, default = FALSE.}
}
\value{
Same object as input, adapted.
}
\description{
Adapts Spanish encoding when using UTF8, deleting accents (tildes) and
turning ñ into n, ç into c, ü into u.
}
\examples{
adapt("Ñora")
adapt("Lledó", tolower = TRUE)

}
