% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changes_groups.R
\name{changes_groups}
\alias{changes_groups}
\title{Group together municipalities that were once merged.}
\usage{
changes_groups(y_start, y_end, prov = NULL, partial_changes = FALSE,
  checks = FALSE)
}
\arguments{
\item{y_start}{Integer, earliest year for which changes are considered}

\item{y_end}{Integer, latest year for which changes are considered}

\item{prov}{If a list of provinces is specified, only municipalities in
  those provinces are returned (Character vector, detault = NULL).
  If NULL, all provinces are included.

  Accepts official names, case-insentitive: Alava, Albacete, Alicante,
  Almeria, Avila, Badajoz, Baleares, Barcelona, Burgos, Caceres, Cadiz,
  Castellon, Ceuta, Ciudad Real, Cordoba, A Coruna, Cuenca, Girona,
  Granada, Guadalajara, Gipuzkoa, Huelva, Huesca, Jaen, Leon, Lleida,
  La Rioja, Lugo, Madrid, Malaga, Melilla, Murcia, Navarra, Ourense,
  Asturias, Palencia, Las Palmas, Pontevedra, Salamanca,
  Santa Cruz de Tenerife, Cantabria, Segovia, Sevilla, Soria, Tarragona,
  Teruel, Toledo, Valencia, Valladolid, Bizkaia, Zamora, Zaragoza.}

\item{partial_changes}{(*In progress*) If TRUE, all municipalities that suffered
  partial changes will also be merged (Logical, default = FALSE).

  Partial changes refers to those cases where only a part of a municipality
  is affected, either because it gets transfered to another municipality or
  because a whole municipality splits and its territories become part of
  different municipalities.}

\item{checks}{If TRUE, checks whether output includes all municipalities
and no municipality is included in two groups (Logical, default = FALSE).
*Note:* this option increases running time considerably.}
}
\value{
A character vector with the municipality codes group together in
  each string, separated by ';'.
}
\description{
Group together municipalities that were once merged.
}
\examples{
ch_euk = changes_groups(1930, 1970,
  prov = c("gipuzkoa", "alava", "bizkaia"), checks = TRUE)
head(ch_euk)
tail(ch_euk)

}
