% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_to_name.R
\name{code_to_name}
\alias{code_to_name}
\title{Returns census names corresponding to municipality codes.}
\usage{
code_to_name(code)
}
\arguments{
\item{code}{Municipality code. If using 6-digit municipalities, use character class.}
}
\value{
Character vector. NA if not found.
}
\description{
Returns census names corresponding to municipality codes.
}
\examples{
  code_to_name(c(15005, 48020))
  code_to_name(c(35001))
  code_to_name(c(35001, "035001"))

}
