% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_to_prov.R
\name{code_to_prov}
\alias{code_to_prov}
\title{Get province name from province code.}
\usage{
code_to_prov(code)
}
\arguments{
\item{code}{Character or integer vector. Valid numbers from 1 to 52.}
}
\value{
Character, province name (official, lower-case).
}
\description{
Get province name from province code.
}
\examples{
code_to_prov(c(1, 2, 3, "33"))

}
