% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prov_to_ccaa.R
\name{prov_to_ccaa}
\alias{prov_to_ccaa}
\title{Get CCAA from province name.}
\usage{
prov_to_ccaa(prov)
}
\arguments{
\item{prov}{Character vector.

  Accepts official names, case-insensitive: Alava, Albacete, Alicante,
  Almeria, Avila, Badajoz, Baleares, Barcelona, Burgos, Caceres, Cadiz,
  Castellon, Ceuta, Ciudad Real, Cordoba, A Coruna, Cuenca, Girona,
  Granada, Guadalajara, Gipuzkoa, Huelva, Huesca, Jaen, Leon, Lleida,
  La Rioja, Lugo, Madrid, Malaga, Melilla, Murcia, Navarra, Ourense,
  Asturias, Palencia, Las Palmas, Pontevedra, Salamanca,
  Santa Cruz de Tenerife, Cantabria, Segovia, Sevilla, Soria, Tarragona,
  Teruel, Toledo, Valencia, Valladolid, Bizkaia, Zamora, Zaragoza.
  Also: Araba, La Coruna, Orense, Gerona, Lerida, Guipuzcoa, Vizcaya.}
}
\value{
Integer, province code. Returns simplified names, not official ones.
}
\description{
Get CCAA from province name.
}
\examples{
prov_to_ccaa(c("Asturias", "asturias", "Lugo", "lleida", "Lerida"))

}
