% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prov_to_code.R
\name{prov_to_code}
\alias{prov_to_code}
\title{Get province code from province name.}
\usage{
prov_to_code(prov)
}
\arguments{
\item{prov}{Character vector.

  Accepts official names, case-insensitive: Alava, Albacete, Alicante,
  Almeria, Avila, Badajoz, Baleares, Barcelona, Burgos, Caceres, Cadiz,
  Castellon, Ceuta, Ciudad Real, Cordoba, A Coruna, Cuenca, Girona,
  Granada, Guadalajara, Gipuzkoa, Huelva, Huesca, Jaen, Leon, Lleida,
  La Rioja, Lugo, Madrid, Malaga, Melilla, Murcia, Navarra, Ourense,
  Asturias, Palencia, Las Palmas, Pontevedra, Salamanca,
  Santa Cruz de Tenerife, Cantabria, Segovia, Sevilla, Soria, Tarragona,
  Teruel, Toledo, Valencia, Valladolid, Bizkaia, Zamora, Zaragoza.
  Also: Araba, La Coruna, Orense, Gerona, Lerida, Guipuzcoa, Vizcaya.}
}
\value{
Integer, province code.
}
\description{
Get province code from province name.
}
\examples{
prov_to_code(c("Asturias", "asturias", "Lugo", "lleida", "Lerida"))

}
