% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_to_code.R
\name{name_to_code}
\alias{name_to_code}
\title{Returns INE codes corresponding to municipality names.}
\usage{
name_to_code(muni, prov = NULL)
}
\arguments{
\item{muni}{Character string or character vector. It returns a character vector of
INE codes equivalent to the input. If not found, it returns NA.}

\item{prov}{Optional, character vector. Include in case of name duplicates.
  If a duplicate is found and no province was specified, it returns an error message.

  Accepts official names, case-insentitive: Alava, Albacete, Alicante,
  Almeria, Avila, Badajoz, Baleares, Barcelona, Burgos, Caceres, Cadiz,
  Castellon, Ceuta, Ciudad Real, Cordoba, A Coruna, Cuenca, Girona,
  Granada, Guadalajara, Gipuzkoa, Huelva, Huesca, Jaen, Leon, Lleida,
  La Rioja, Lugo, Madrid, Malaga, Melilla, Murcia, Navarra, Ourense,
  Asturias, Palencia, Las Palmas, Pontevedra, Salamanca,
  Santa Cruz de Tenerife, Cantabria, Segovia, Sevilla, Soria, Tarragona,
  Teruel, Toledo, Valencia, Valladolid, Bizkaia, Zamora, Zaragoza.}
}
\value{
Character vector equivalent to the input. NA if not found.
}
\description{
Returns INE codes corresponding to municipality names.
}
\examples{
  # name_to_code("Tapia") # Returns an error
  # name_to_code(c("Tapia", "Castropol")) # Returns an error
  name_to_code("Tapia de Casariego")
  name_to_code(c("Tapia", "Castropol"), rep("Asturias", 2))


}
